+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Fixed ACL username support when using Predis and array syntax
 - Made admin-bar script more robust
 - Improved CloudLinux detection
+- Prevent LiteSpeed Cache from overwriting drop-in
 
 ## 2.4.1